updates to slow mode room (previously room) stuff
